pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-novel-generator"
version = "1.0.0"
description = "一款基于大语言模型的多功能小说生成器，助您高效创作逻辑严谨、设定统一的长篇故事"
readme = "README.md"
authors = [
    {name = "YILING0013", email = "example@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "azure-ai-inference==1.0.0b9",
    "azure-core==1.35.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "build==1.3.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "chromadb==1.0.20",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "customtkinter==5.2.2",
    "darkdetect==0.8.0",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "durationpy==0.10",
    "filelock==3.19.1",
    "flatbuffers==25.2.10",
    "frozenlist==1.7.0",
    "fsspec==2025.9.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.181.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.30.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.4",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib_metadata==8.7.0",
    "importlib_resources==6.5.2",
    "isodate==0.7.2",
    "Jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "keybert==0.9.0",
    "kubernetes==33.1.0",
    "langchain==0.3.27",
    "langchain-chroma==0.2.5",
    "langchain-community==0.3.29",
    "langchain-core==0.3.75",
    "langchain-openai==0.3.32",
    "langchain-text-splitters==0.3.11",
    "langsmith==0.4.25",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "mypy_extensions==1.1.0",
    "networkx==3.5",
    "nltk==3.9.1",
    "numpy==2.3.2",
    "oauthlib==3.3.1",
    "onnxruntime==1.22.1",
    "openai==1.106.1",
    "opentelemetry-api==1.36.0",
    "opentelemetry-exporter-otlp-proto-common==1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.36.0",
    "opentelemetry-proto==1.36.0",
    "opentelemetry-sdk==1.36.0",
    "opentelemetry-semantic-conventions==0.57b0",
    "orjson==3.11.3",
    "overrides==7.7.0",
    "packaging==25.0",
    "pillow==11.3.0",
    "posthog==5.4.0",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pybase64==1.4.2",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pydantic_core==2.33.2",
    "Pygments==2.19.2",
    "pyparsing==3.2.3",
    "PyPika==0.48.9",
    "pyproject_hooks==1.2.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "regex==2025.9.1",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "sentence-transformers==5.1.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.43",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.11.0",
    "tokenizers==0.22.0",
    "torch==2.8.0",
    "tqdm==4.67.1",
    "transformers==4.56.1",
    "typer==0.17.4",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.15.0",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchfiles==1.1.0",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wheel==0.45.1",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.24.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
]

[project.scripts]
ai-novel-generator = "main:main"

[project.gui-scripts]
ai-novel-generator-gui = "main:main"

[project.urls]
Homepage = "https://github.com/YILING0013/AI_NovelGenerator"
Repository = "https://github.com/YILING0013/AI_NovelGenerator"
Issues = "https://github.com/YILING0013/AI_NovelGenerator/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["novel_generator*", "ui*"]

[tool.setuptools.package-data]
"*" = ["*.json"]